refactor(eidos): extract knowledge::entity submodule (#2642)
Move Entity, Relationship, EmbeddedChunk, RecallResult to
knowledge/entity.rs. Add explicit id imports to test file since
EntityId and EmbeddingId are no longer leaked through super::*.
